packs/st2_chatops_misc/requirements.txt
python-libxml2
python-httplib2
